refactor: extract chat helpers and save messages logic into separate hooks for improved modularity
